+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v0.1.3 (2021/10/04)
 
 - Class name changed from tidyframe to tibble
+- Features
+  - tibble::bind_rows
+  - tibble::bind_cols
 
 ## v0.1.0 (2021/10/02)